+++ src/support/contracts/highlevel/datastructures/firehub.Collection.php
@@ -31,6 +31,30 @@ use FireHub\Core\Support\Contracts\ArrayAccessible;
 interface Collection extends Dynamic, ArrayAccessible {
 
     /**
+     * ### Get first item from collection
+     * @since 1.0.0
+     *
+     * @param null|callable(TValue=, TKey=):bool $callback [optional] <p>
+     * If callback is used, the method will return the first item that passes the truth test.
+     * </p>
+     *
+     * @return null|TValue First item from a collection.
+     */
+    public function first (?callable $callback = null):mixed;
+
+    /**
+     * ### Get last item from collection
+     * @since 1.0.0
+     *
+     * @param null|callable(TValue=, TKey=):bool $callback [optional] <p>
+     * If callback is used, the method will return the last item that passes the truth test.
+     * </p>
+     *
+     * @return null|TValue Last item from a collection.
+     */
+    public function last (?callable $callback = null):mixed;
+
+    /**
      * ### Get collection items as an array
      * @since 1.0.0
      *