+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased](https://github.com/The-FireHub-Project/Core/compare/v0.1.6...develop-pre-alpha-m1)
 
+### Added
+- Create Arr helper functions, first and last ([#18](https://github.com/The-FireHub-Project/Core/issues/18), [1019f93](https://github.com/The-FireHub-Project/Core/pull/4/commits/1019f93))
+
 # [v0.1.6](https://github.com/The-FireHub-Project/Core/compare/v0.1.5...v0.1.6) – (2025-03-18)
 
 ### Added